Moved words modification methods into separated class
